# src/swagger-local.json
{
  "swagger": "2.0",
  "info": {
    "title": "Spruce Moose API",
    "description": "🌲 SPRUCE MOOSE 🌲",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Create new user AFTER successful OAuth login",
            "schema": {
              "$ref": "#/definitions/newUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Get a user by ID. Frontend should never have to use this, this is for debugging only",
            "schema": {
              "$ref": "#/definitions/newUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/plant/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scientificName": {
                  "example": "any"
                },
                "commonName": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "imgUrl": {
                  "example": "any"
                },
                "stockQty": {
                  "example": "any"
                },
                "orderQty": {
                  "example": "any"
                },
                "wasteQty": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/plant/{plantId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "plantId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "plantId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "scientificName": {
                  "example": "any"
                },
                "commonName": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "imgUrl": {
                  "example": "any"
                },
                "stockQty": {
                  "example": "any"
                },
                "orderQty": {
                  "example": "any"
                },
                "wasteQty": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "newUser": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Joe"
        },
        "lastName": {
          "type": "string",
          "example": "Momma"
        }
      },
      "required": ["firstName", "lastName"]
    },
    "error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Something went wrong."
        },
        "error": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "details"
            }
          }
        }
      },
      "required": ["message"]
    }
  }
}